refactor: extract helpers
Move manifest assertion helpers to own module so they can be reused
